Adicionado função para pegar tecla do usuário e preparativos para fazer o tabuleiro começar a se mexer conforme a entrada do usuário.
