+++ bloco.h
@@ -1,5 +1,7 @@
 // PROTÓTIPOS DE STRUCTS -----------------------
 
+
+
 // -------------------------------------------
 
 // PROTÓTIPOS DE FUNÇÕES ----------------------
